requirements.txt
pandas==1.5.3
geopandas==0.13.2
shapely==2.0.7
pyproj==3.6.1
fiona==1.9.6
pyarrow==15.0.2
streamlit>=1.40,<2
plotly>=5.20,<6
python-dotenv>=1.0,<2
pytest>=8.0,<9
